requirements.txt
asgiref==3.5.2
Django==4.0.6
django-cors-headers==3.13.0
djangorestframework==3.13.1
djongo==1.3.6
dnspython==2.2.1
mongoengine==0.24.1
pymongo==4.1.1
pytz==2022.1
sqlparse==0.2.4
